# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Catalog API
  description: Node.js Catalog RESTful API
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Auth server
  - url: http://localhost:8082
    description: Users server
  - url: http://localhost:8083
    description: Products server
...
paths:
...
  /auth/refresh/{refreshToken}:
    get:
      tags: 
        - Authentication
      description: Generates new Access Token
      operationId: refresh
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            $ref: "#/components/schemas/Authorization"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  description: Refresh Token
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImJydW5vQGVtYWlsLmNvbSIsInBhc3N3b3JkIjoiMTIzNDUiLCJpYXQiOjE2MjI5MjUyNzF9._-5EiwijyMU1y49adL1hRTOgQx-xipBdnun2Fym3VWw
              required:
                - refreshToken
      responses:
        "200":
          description: New Access Token Generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    description: Access Token
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImJydW5vQGVtYWlsLmNvbSIsInBhc3N3b3JkIjoiMTIzNDUiLCJpYXQiOjE2MjI5MjUyNzEsImV4cCI6MTYyMjkyNTU3MX0.0I5gFrVjyUjdgg_NkMwdwk6ZHux4n8JLqgia1CERrbM
                  refreshToken:
                    description: Refresh Token
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImJydW5vQGVtYWlsLmNvbSIsInBhc3N3b3JkIjoiMTIzNDUiLCJpYXQiOjE2MjI5MjUyNzF9._-5EiwijyMU1y49adL1hRTOgQx-xipBdnun2Fym3VWw
                required:
                  - accessToken
                  - refreshToken
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/400Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/401Response"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500Response"
        default:
          description: Some other error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /isAdmin:
    get:
      tags: 
        - Authentication
      description: Search users
      operationId: isAdmin
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  isAdmin:
                    description: Is user admin?
                    type: boolean
                    example: true
                required:
                  - isAdmin
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403Response"
...
components:
  schemas:
    Authorization:
      type: object
      properties:
        Authorization:
          description: Access Token
          type: string
          example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImJydW5vQGVtYWlsLmNvbSIsInBhc3N3b3JkIjoiMTIzNDUiLCJpYXQiOjE2MjI5MjUyNzEsImV4cCI6MTYyMjkyNTU3MX0.0I5gFrVjyUjdgg_NkMwdwk6ZHux4n8JLqgia1CERrbM
...
    400Response:
      type: object
      properties:
        error:
          description: Error message
          type: string
          example: Bad Request
      required:
        - error
    401Response:
      type: object
      properties:
        error:
          description: Error message
          type: string
          example: Unauthorized
      required:
        - error
    403Response:
      type: object
      properties:
        error:
          description: Error message
          type: string
          example: Forbidden
      required:
        - error
...
    500Response:
      type: object
      properties:
        error:
          description: Error message
          type: string
          example: Internal Server Error
      required:
        - error
    ErrorResponse:
      type: object
      properties:
        error:
          description: Error message
          type: string
          example: Error message
      required:
        - error